refactor: move insight fund flow types
